proto/types: use fixed-length 64-bit integers for rounds in CanonicalVoteExtension

+++ types/canonical.go
@@ -72,7 +72,7 @@ func CanonicalizeVoteExtension(chainID string, vote *tmproto.Vote) tmproto.Canon
 	return tmproto.CanonicalVoteExtension{
 		Extension: vote.Extension,
 		Height:    vote.Height,
-		Round:     vote.Round,
+		Round:     int64(vote.Round),
 		ChainId:   chainID,
 	}
 }
